:alien: 設定を別ファイルにする
